+++ src/RgbImage.cpp
@@ -7,10 +7,10 @@
 
 void RgbImage::to_gray_image(GrayImage &image)
 {
-    auto matrix = std::make_shared<ublas::matrix<uint>>(this->height, this->width);
-    ublas::matrix<uint> b = (*get_chunk(1));
-    ublas::matrix<uint> g = (*get_chunk(2));
-    ublas::matrix<uint> r = (*get_chunk(3));
+    auto matrix = std::make_shared<ublas::matrix<uchar>>(this->height, this->width);
+    ublas::matrix<uchar> b = (*get_chunk(1));
+    ublas::matrix<uchar> g = (*get_chunk(2));
+    ublas::matrix<uchar> r = (*get_chunk(3));
     for (int i = 0; i < this->height; i++) {
         for (int j = 0; j < this->width; j++) {
             (*matrix.get())(i, j) = ceil(r(i,j)*0.3+g(i,j)*0.59+b(i,j)*0.11);
